2020/11/13

+++ main.py
@@ -113,22 +113,23 @@ async def on_message(msg):
 
 @bot.command()
 async def 十連抽(msg):
-    embed = discord.Embed()
-    toImage = Image.new('RGBA',(600,240),color="white")
 
-    #pic = ["https://i.imgur.com/Nx72jjy.png","https://i.imgur.com/KsCi7Xa.png","https://i.imgur.com/3tGzRNA.png","https://yys.res.netease.com/pc/zt/20161108171335/data/shishen/217.png?v1"]
+    rolledchannellist = list(jdata_1["RolledChannel"])
+    
+    if msg.channel.id in rolledchannellist:
+        embed = discord.Embed()
+        toImage = Image.new('RGBA',(600,240),color="white")
 
-    for i in range(3):
-        response = req.get(jdata_5["pic"][i])
-        fromImge = Image.open(BytesIO(response.content))
-        # loc = ((i % 2) * 200, (int(i/2) * 200))
-        loc = ((int(i/2) * 120), (i % 2) * 120)
-        toImage.paste(fromImge, loc)
+        for i in range(3):
+            response = req.get(jdata_5["pic"][i])
+            fromImge = Image.open(BytesIO(response.content))
+            loc = ((int(i/2) * 120), (i % 2) * 120)
+            toImage.paste(fromImge, loc)
 
-    toImage.save('01.png')
-    file = discord.File("01.png", filename="image.png")
-    embed.set_image(url="attachment://image.png")
-    await msg.channel.send(file=file, embed = embed)
-    os.remove("01.png")
+        toImage.save('01.png')
+        file = discord.File("01.png", filename="image.png")
+        embed.set_image(url="attachment://image.png")
+        await msg.channel.send(file=file, embed = embed)
+        os.remove("01.png")
 
 bot.run(os.environ['TOKEN'])
